+++ packages/react/src/runtime/noop-handle.ts
@@ -12,6 +12,7 @@ export const NOOP_HANDLE: DevivaDrawHandle = {
   exportToPng: () => Promise.reject(new Error("deviva-draw: not mounted yet")),
   exportToSvg: () => "<svg xmlns=\"http://www.w3.org/2000/svg\" width=\"0\" height=\"0\"></svg>",
   resetScene: () => {},
+  newDocument: () => {},
   loadSceneDocument: () => false,
   undo: () => {},
   redo: () => {},